Revert "Creating a release 0.1.1 (#32) (#33)"
This reverts commit ae497a376b9808142d57efc63fbe322a2f078694.
