+++ Backend/package.json
@@ -4,8 +4,7 @@
   "description": "",
   "main": "server.js",
   "scripts": {
-    "test": "jest --runInBand --forceExit",
-    "start": "nodemon ./src/server.js"
+    "test": "jest --runInBand --forceExit"
   },
   "author": "",
   "license": "MIT",
@@ -13,23 +12,19 @@
     "@prisma/client": "^5.2.0",
     "bcrypt": "^5.1.1",
     "body-parser": "^1.20.2",
-    "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
-    "express-session": "^1.17.3",
     "googleapis": "^105.0.0",
     "joi": "^17.10.1",
     "nodemailer": "^6.9.4",
     "passport": "^0.6.0",
-    "passport-local": "^1.0.0",
     "prisma": "^5.2.0",
     "rand-token": "^1.0.1",
     "zod": "^3.22.2"
   },
   "devDependencies": {
     "jest": "^29.7.0",
-    "jest-fetch-mock": "^3.0.3",
     "nodemon": "^3.0.1",
     "supertest": "^6.3.3"
   }
