--- Backend/src/components/ensure_authenticated.js
@@ -1,7 +1,0 @@
-module.exports = function ensureAuthenticated(req, res, next) {
-    if (req.isAuthenticated()) {
-        return next();
-    } else {
-        res.status(401).json({"message": "Usuário não logado"});
-    }
-}
