+++ Backend/src/email_service.js
@@ -1,3 +1,4 @@
+// valeu indiano <3 te amo
 require('dotenv').config();
 const nodemailer = require('nodemailer');
 const { google } = require('googleapis');
