--- Backend/src/tests/register_user.test.js
@@ -1,81 +1,0 @@
-const Joi = require("joi");
-const validationSchema = require("../validation/register_user");
-
-describe("Validation Schema", () => {
-    it("deve permitir um objeto com valores válidos", () => {
-        const validObject = {
-            first_name: "John",
-            last_name: "Doe",
-            email: "john@example.com",
-            password: "secure123",
-        };
-    
-        const result = validationSchema.validate(validObject);
-    
-        expect(result.error).toBeUndefined();
-        });
-
-    it("deve retornar um erro para um objeto sem first_name", () => {
-        const invalidObject = {
-            last_name: "Doe",
-            email: "john@example.com",
-            password: "secure123",
-        };
-
-        const result = validationSchema.validate(invalidObject);
-
-        expect(result.error).not.toBeNull();
-    });
-
-    it("deve retornar um erro para um objeto com last_name invalido", () => {
-        const invalidObject = {
-            first_name: "John",
-            last_name: "D",
-            email: "john@example.com",
-            password: "secure123",
-        };
-
-        const result = validationSchema.validate(invalidObject);
-
-        expect(result.error).not.toBeNull();
-    });
-
-    it("deve retornar um erro para um objeto com email invalido", () => {
-        const invalidObject = {
-            first_name: "John",
-            last_name: "Doe",
-            email: "john",
-            password: "secure123",
-        };
-
-        const result = validationSchema.validate(invalidObject);
-
-        expect(result.error).not.toBeNull();
-    });
-
-    it("deve retornar um erro para um objeto com password invalido", () => {
-        const invalidObject = {
-            first_name: "John",
-            last_name: "Doe",
-            email: "john@example.com",
-            password: "se",
-        };
-
-        const result = validationSchema.validate(invalidObject);
-
-        expect(result.error).not.toBeNull();
-    });
-
-    it("deve retornar um erro para um objeto com first_name muito curto", () => {
-        const invalidObject = {
-            first_name: "Jo",
-            last_name: "Doe",
-            email: "john@example.com",
-            password: "secure123",
-        };
-
-        const result = validationSchema.validate(invalidObject);
-
-        expect(result.error).not.toBeNull();
-    });
-});
