+++ Backend/src/validation/confirm_user.js
@@ -1,9 +1,6 @@
 const Joi = require("joi");
 
 module.exports = Joi.object({
-    username: Joi.string()
-        .required(),
-        
     email: Joi.string()
         .email()
         .required(),
