--- Backend/src/validation/create_newsletter.js
@@ -1,13 +1,0 @@
-const Joi = require("joi");
-
-module.exports = Joi.object({
-    name: Joi.string()
-        .max(30)
-        .min(5)
-        .required(),
-    
-    description: Joi.string()
-        .max(200)
-        .min(5)
-        .required()
-})
