--- Backend/src/validation/login_user.js
@@ -1,9 +1,0 @@
-const Joi = require('joi');
-
-module.exports = Joi.object({
-    username: Joi.string()
-        .required(),
-
-    password: Joi.string()
-        .required()
-})
