--- Backend/src/validation/newsletters_from_user.js
@@ -1,3 +1,0 @@
-const Joi = require('joi');
-
-module.exports = Joi.string().required()
