+++ Backend/src/validation/register_user.js
@@ -1,11 +1,6 @@
 const Joi = require("joi");
 
 module.exports = Joi.object({
-    username: Joi.string()
-        .min(6)
-        .max(20)
-        .required(),
-        
     first_name: Joi.string()
         .min(3)
         .max(30)
