--- Frontend/src/pages/authenticationRoot.tsx
@@ -1,11 +1,0 @@
-import { Outlet } from "react-router-dom"
-
-function AuthenticationRoot() {
-  return (
-    <div className="container">
-      <Outlet />
-    </div>
-  );
-}
-
-export default AuthenticationRoot;
