+++ Frontend/src/pages/emailAuthenticationPage.tsx
@@ -2,7 +2,7 @@ import { useNavigate } from "react-router-dom";
 import { ConfirmButton } from "../components/ConfirmButton";
 import { Input } from "../components/Input";
 import "../styles/emailAuthentication.css";
-import { FormEvent, useEffect, useState } from "react";
+import { FormEvent, useState } from "react";
 import "../styles/dialog.css"
 import api from "../services/api";
 
@@ -14,41 +14,27 @@ export function EmailAuthentication() {
   const navigate = useNavigate();
 
   const [tokenCode, setTokenCode] = useState("");
+  const [email, setEmail] = useState("");
+  const [emailSend, setEmailSend] = useState(false);
   const [dialogOpen, setDialogOpen] = useState(false);
   const [errorText, setErrorText] = useState("");
   const [errorTextSwitcher, setErrorTextSwitcher] = useState(false);
 
-  useEffect(() => {
-    const username = localStorage.getItem("username")
-    const name = localStorage.getItem("first_name")
-    const lastName = localStorage.getItem("last_name")
-    const email = localStorage.getItem("email")
-    const validate = localStorage.getItem("validate")
-    if(validate) {
-      navigate("/")
-    } else if(!email || !username || !name || !lastName) {
-      navigate("/authentication/singup")
-    }
-  }, [navigate])
+  const handleEmailAdd = (event: FormEvent) => {
+    event.preventDefault();
+
+    setEmailSend(true);
+  }
 
   const handleConfirmation = async (event: FormEvent) => {
     event.preventDefault();
 
-    const username = localStorage.getItem("username")
-    const email = localStorage.getItem("email")
-
     api.post('confirm_user', {
-      "username": username,
       "email": email,
       "token_confirmation": tokenCode
-    },
-    {
-      withCredentials: true
-    }
-    ).then(async () => {
+    }).then(async () => {
       if(!dialogOpen) {
         setDialogOpen(true);
-        localStorage.setItem("validate", "validated user")
         await delay(1500);
         navigate("/");
       } else {
@@ -56,37 +42,23 @@ export function EmailAuthentication() {
       }
     }).catch(async (error) => {
       if(error.response.status === 404) {
-        setErrorText("O email informado não existe, refaça o cadastro");
-        localStorage.removeItem("username")
-        localStorage.removeItem("first_name")
-        localStorage.removeItem("last_name")
-        localStorage.removeItem("email")
-        localStorage.removeItem("validate")
-        await delay(2500);
+        setErrorText("O email informado não existe");
+        await delay(2000);
+        setEmailSend(false);
       }
       else if(error.response.status === 400) {
         if(error.response.data.message.includes("token")) {
           setErrorText("Digite o código enviado no email");
         }
         else if(error.response.data.message.includes("email")) {
-          setErrorText("Ocorreu um erro, por favor refaça o cadastro");
-          localStorage.removeItem("username")
-          localStorage.removeItem("first_name")
-          localStorage.removeItem("last_name")
-          localStorage.removeItem("email")
-          localStorage.removeItem("validate")
-          await delay(2500);
+          setErrorText("O campo email não pode estar vazio");
+          await delay(2000);
+          setEmailSend(false);
         }
       }
       else if(error.response.status === 401) {
         if(error.response.data.message.includes("expirado")) {
           setErrorText("O seu código de confirmação expirou");
-          localStorage.removeItem("username")
-          localStorage.removeItem("first_name")
-          localStorage.removeItem("last_name")
-          localStorage.removeItem("email")
-          localStorage.removeItem("validate")
-          await delay(2000)
         }
         else if(error.response.data.message.includes("coincidem")) {
           setErrorText("Código incorreto, tente novamente");
@@ -94,9 +66,6 @@ export function EmailAuthentication() {
       }
       else if(error.response.status === 409) {
         setErrorText("Email de usuário já confirmado")
-        localStorage.setItem("validate", "validated user")
-        await delay(1000)
-        navigate("/")
       }
       setErrorTextSwitcher(true);
     })
@@ -104,34 +73,56 @@ export function EmailAuthentication() {
 
   return (
       <>
-        <div className="App">              
-          <form className='email-code-form' onSubmit={handleConfirmation}>
-            <h1>Informe o código de confirmação</h1>
-            <Input
-              type="text"
-              name="tokenCode"
-              id="tokenCode"
-              placeholder='Digite aqui o código'
-              value={tokenCode}
-              onChange={
-                (event) => {
-                  setTokenCode(event.target.value)
-                }
-              }
-            />
-            {errorTextSwitcher ?
-              <p className="error-text">{errorText}</p>
+        <div className="container">
+            { !emailSend ?
+              <form className='email-code-form' onSubmit={handleEmailAdd}>
+                <h1>Informe seu email</h1>
+                <Input
+                  type="email"
+                  name="email"
+                  id="email"
+                  placeholder='Digite seu email aqui'
+                  value={email}
+                  onChange={
+                    (event) => {
+                      setEmail(event.target.value)
+                    }
+                  }
+                  required
+                />
+                <ConfirmButton type="submit">Continuar</ConfirmButton>
+              </form>
               :
-              <p>Enviamos um código no email cadastrado.</p>
+              <form className='email-code-form' onSubmit={handleConfirmation}>
+                <h1>Informe o código de confirmação</h1>
+                <Input
+                  type="text"
+                  name="tokenCode"
+                  id="tokenCode"
+                  placeholder='Digite aqui o código'
+                  value={tokenCode}
+                  onChange={
+                    (event) => {
+                      setTokenCode(event.target.value)
+                    }
+                  }
+                />
+                {errorTextSwitcher ?
+                  <p className="error-text">{errorText}</p>
+                  :
+                  <p>Enviamos um código no email cadastrado.</p>
+                }
+                <ConfirmButton type="submit" onSubmit={handleConfirmation}>Continuar</ConfirmButton>
+              </form>
             }
-            <ConfirmButton type="submit" onSubmit={handleConfirmation}>Continuar</ConfirmButton>
-          </form>
         </div>
 
-        {dialogOpen &&
+        {dialogOpen ?
           <dialog className="dialog">
             <h1>Email confirmado</h1>
           </dialog>
+          :
+          null
         }
     </>
   )
