+++ Frontend/src/pages/error404Page.tsx
@@ -1,8 +1,7 @@
 export function Error404Page() {
   return (
-    <div className="App" style={{flexDirection: "column"}}>
-      <h1 style={{color: "white"}}>Error 404!</h1>
-      <p style={{color: "white"}}>Página não encontrada</p>
+    <div>
+      <h3>Error 404!</h3>
     </div>
   )
 }
