+++ Frontend/src/services/api.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 
 const api = axios.create({
-  baseURL: process.env.REACT_APP_API_URL
+  baseURL: 'http://10.0.0.124:3333'
 });
 
 export default api;
