--- Frontend/src/styles/createNewsletter.css
@@ -1,61 +1,0 @@
-.container-new-newsletter {
-  height: 100%;
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.form-new-newsletter {
-  height: 100%;
-  width: 60%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.form-new-newsletter h1 {
-  margin-top: 1rem;
-}
-
-.label-new-newsletter {
-  margin-top: 3%;
-  font-size: 1rem;
-  color: white;
-}
-
-.name-new-newsletter {
-  width: 60%;
-  margin: 0;
-  margin-top: 1%;
-  border-bottom: .2em solid white;
-  background-color: white;
-}
-
-.textarea-new-newsletter {
-  margin-top: 1%;
-  width: 60%;
-  background: white;
-  text-align: center;
-	outline: none;
-	border-bottom: .2em solid white;
-	color: black;
-	font-size: 70%;
-	padding: 1em;
-	border-radius: 2em;
-	letter-spacing: 0.01em;
-  resize: none;
-}
-
-.button-new-newsletter {
-  margin: .5rem 0 .5rem 0;
-  width: 25%;
-  color: white;
-  border-color: white;
-}
-
-.button-new-newsletter:hover {
-  background-color: white;
-  color: black;
-}
