--- Frontend/src/styles/footer.css
@@ -1,21 +1,0 @@
-.app-footer {
-    display: flex;
-    width: 100%;
-    height: 15%;
-    background-color: #0D0106;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-}
-
-.app-footer span {
-    font-size: .7rem;
-}
-
-.app-footer :last-child {
-    cursor: pointer;
-}
-
-.app-footer :last-child:hover {
-    color: #CCC;
-}
