--- Frontend/src/styles/header.css
@@ -1,48 +1,0 @@
-.app-header {
-    display: flex;
-    width: 100%;
-    height: 15%;
-    background-color: #6B7FD7;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.app-header img {
-    width: 1.7rem;
-    height: 1.7rem;
-    border-radius: 10rem;
-}
-
-.app-name {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-left: 1rem;
-}
-
-.app-name span {
-    margin-left: .4rem;
-    font-size: 1rem;
-    font-weight: bolder;
-    cursor: pointer;
-}
-
-.menu-items {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: .8rem;
-    margin-right: 1rem;
-}
-
-.menu-items span {
-    text-decoration: none;
-    color: white;
-    margin-left: .5rem;
-    font-size: .9rem;
-    cursor: pointer;
-}
-
-.menu-items span:hover {
-    color: #CCC;
-}
