--- Frontend/src/styles/homePage.css
@@ -1,77 +1,0 @@
-.home-container {
-    height: 100%;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-self: center;
-    flex-direction: column;
-    background-image: url("../images/home-page.jpg");
-    background-size: cover;
-    background-repeat: none;
-}
-
-.text-container {
-    height: 100%;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-self: center;
-}
-
-.button-container {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 20%;
-    width: 100%;
-    background-color: rgba(0, 0, 0, 0.50);
-    padding-bottom: 10%;
-}
-
-.create-text {
-    display: flex;
-    align-items: start;
-    justify-content: start;
-    flex-direction: column;
-    width: 50vw;
-    height: 100%;
-    padding-left: 5%;
-    padding-right: 10%;
-    padding-top: 10%;
-    background-color: rgba(0, 0, 0, 0.50);
-}
-
-.create-text span {
-    font-size: .7rem;
-}
-
-.singup-text span {
-    font-size: .7rem;
-}
-
-.singup-text {
-    display: flex;
-    align-items: start;
-    justify-content: start;
-    flex-direction: column;
-    width: 50vw;
-    height: 100%;
-    padding-left: 5%;
-    padding-right: 10%;
-    padding-top: 10%;
-    background-color: rgba(0, 0, 0, 0.50);
-}
-
-.button-container button {
-    width: 15%;
-    color: white;
-    border-color: white;
-    font-weight: bold;
-}
-
-.button-container button:hover {
-    color: #0D0106;
-    border-color: white;
-    background-color: white;
-    cursor: pointer;
-}
