--- Frontend/src/styles/homePageLogged.css
@@ -1,55 +1,0 @@
-.home-logged-container {
-  height: 100%;
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-self: center;
-  background-image: url("../images/home-page-logged.jpg");
-  background-size: cover;
-  background-repeat: no-repeat;
-}
-
-.new-newsletter-container {
-  display: flex;
-  align-items: start;
-  justify-content: start;
-  flex-direction: column;
-  width: 50vw;
-  height: 100%;
-  padding-left: 10%;
-  padding-right: 10%;
-  padding-top: 10%;
-  background-color: rgba(0, 0, 0, 0.50);
-}
-
-.subscribe-newsletter-container {
-  display: flex;
-  align-items: start;
-  justify-content: start;
-  flex-direction: column;
-  width: 50vw;
-  height: 100%;
-  padding-left: 10%;
-  padding-right: 10%;
-  padding-top: 10%;
-  background-color: rgba(0, 0, 0, 0.50);
-}
-
-.home-logged-container p {
-  font-size: .8rem;
-  color: white;
-}
-
-.home-logged-container button {
-  width: 50%;
-  color: white;
-  border-color: white;
-  font-weight: bold;
-}
-
-.button-newsletter-container:hover {
-  color: #0D0106;
-  border-color: white;
-  background-color: white;
-  cursor: pointer;
-}
