+++ Frontend/src/styles/input.css
@@ -1,12 +1,12 @@
 input {
 	margin-top: 2%;
-  width: 70%;
+  width: 75%;
   background: transparent;
   text-align: center;
 	outline: none;
 	border-bottom: .2em solid #0D1B1E;
 	color: #0D1B1E;
-	font-size: 70%;
+	font-size: 75%;
 	padding: .7em;
 	border-radius: 3em;
 	letter-spacing: 0.01em;
