--- Frontend/src/styles/searchNewsletter.css
@@ -1,66 +1,0 @@
-.container-search-newsletter {
-  height: 100%;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: start;
-}
-
-.wrapper-input-search-newsletter {
-  border: 1px white solid;
-  border-radius: 3rem;
-  display: flex;
-  align-items: center;
-  justify-content: space-evenly;
-  width: 50%;
-  height: 10%;
-  margin-top: 1rem;
-  color: white;
-}
-
-.input-search-newsletter {
-  border: none;
-  width: 90%;
-  margin: 0;
-  font-size: 100%;
-  color: white;
-}
-
-.result-items-container-search-newsletter {
-  width: 90%;
-  padding: 1rem;
-  border-radius: 1rem;
-  margin-top: 1rem;
-  background-color: #555358;
-}
-
-.header-search-newsletter {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.title-item-search-newsletter {
-  color: whitesmoke;
-}
-
-.button-search-newsletter {
-  margin: 0;
-  width: 13%;
-  padding: .6rem;
-  border-color: whitesmoke;
-  color: whitesmoke;
-  display: flex;
-  justify-content: center;
-}
-
-.button-search-newsletter:hover {
-  border-color: black;
-}
-
-.description-item-search-newsletter {
-  color: whitesmoke;
-  margin-top: 1rem;
-  font-size: .8rem;
-}
