--- Frontend/src/styles/sidebar.css
@@ -1,84 +1,0 @@
-.sidebar-container {
-  display: flex;
-  align-items: end;
-  justify-content: start;
-  flex-direction: column;
-  background-color: #0D1B1E;
-  position: fixed;
-  top: 0;
-  right: 0;
-  height: 100vh;
-  width: 20vw;
-}
-
-.sidebar-container span {
-  margin: 0;
-}
-
-.sidebar-container span:hover {
-  color: white;
-}
-
-.close-button {
-  color: white;
-}
-
-.logout-button {
-  color: white;
-}
-
-.sidebar-buttons {
-  width: 100%;
-  margin-bottom: 1rem;
-  margin-top: 1rem;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.logout {
-  display: flex;
-  align-items: center;
-  border-radius: .4rem;
-  padding: .4rem;
-}
-
-.logout:hover {
-  background-color: #779197;
-}
-
-.logout span {
-  margin-left: .4rem;
-}
-
-.sidebar-items {
-  display: flex;
-  height: 90%;
-  flex-direction: column;
-  align-items: end;
-  justify-content: start;
-  margin-right: 1rem;
-}
-
-.items-elements {
-  display: flex;
-  align-items: center;
-  justify-content: end;
-  margin-bottom: .8rem;
-  width: 100%;
-  border-radius: .4rem;
-  padding: .4rem;
-}
-
-.items-elements:hover {
-  background-color: #779197;
-}
-
-.items-elements span {
-  font-size: 1rem;
-}
-
-.sidebar-items-icons {
-  color: white;
-  margin-left: 1rem;
-}
