+++ Frontend/src/styles/singUp.css
@@ -1,10 +1,13 @@
-.container {
+.App {
   width: 100vw;
   height: 100vh;
   display: flex;
   align-items: center;
   justify-content: center;
-  flex-direction: column;
+  /* min-height: 100vh; */
+  display: flex;
+  justify-content: center;
+  align-items: center;
   background: #0D1B1E;
 }
 
@@ -13,12 +16,8 @@
   align-items: center;
   justify-content: center;
   text-align: center;
-  height: 80vh;
-  width: 70vw;
-}
-
-.small {
-  height: 50vh;
+  height: 65vh;
+  width: 60vw;
 }
 
 .container-text {
@@ -29,16 +28,7 @@
   height: 100%;
   border-radius: .7rem 0rem 0rem .7rem;
   align-items: center;
-  justify-content: space-between;
-}
-
-.container-text-main {
-  height: 100%;
-  display: flex;
-  align-items: center;
   justify-content: center;
-  flex-direction: column;
-  margin-bottom: 5%;
 }
 
 .container-text button {
@@ -106,28 +96,4 @@ a {
   font-size: .8em;
   margin-top: 2%;
   color: rgb(255, 45, 45);
-}
-
-.feedback {
-  font-size: .6em;
-  margin-top: 1%;
-  color: rgb(255, 45, 45);
-}
-
-.back-button-container {
-  margin-top: 1rem;
-  margin-left: 1rem;
-  align-self: self-start;
-  border-radius: 3rem;
-}
-
-.back-button {
-  color: white;
-  height: 1.5rem;
-  width: 2rem;
-  margin-top: .2rem;
-}
-
-.back-button-container:hover {
-  background-color: rgb(0, 0, 187);
 }
