+++ Frontend/src/tests/App.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import App from '../Root';
+import App from '../App';
 
 test('renders learn react link', () => {
   render(<App />);
